Split monolithic grid_cron.py daemon into a thin sbin/grid_cron.py wrapper with all helper functions in mig/lib/cron.py .
First fundamental splitting here. Still needs various clean ups and unit test
coverage to follow the guidelines for migrating to FHSesque loations.
